refactor(post): restructure request classes into dto package
- 요청 클래스 파일들을 `dto` 전용 패키지로 이동
- 파일 구조 개선을 통해 코드 가독성과 유지보수성 향상
- 관련 import 경로를 모두 업데이트하여 일관된 구조 적용
