+++ src/main/java/dooya/see/domain/post/dto/CommentCreateRequest.java
@@ -1,4 +1,4 @@
-package dooya.see.domain.post;
+package dooya.see.domain.post.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
