+++ src/main/java/dooya/see/domain/post/dto/PostSearchRequest.java
@@ -1,5 +1,7 @@
-package dooya.see.domain.post;
+package dooya.see.domain.post.dto;
 
+import dooya.see.domain.post.PostCategory;
+import dooya.see.domain.post.PostStatus;
 import lombok.Builder;
 
 import java.time.LocalDateTime;
